+++ src/shared/ui/index.tsx
@@ -50,52 +50,52 @@ export const Textarea = forwardRef(({ className, ...props }, ref) => {
 })
 Textarea.displayName = "Textarea"
 
-// 선택 컴포넌트
-export const Select = SelectPrimitive.Root
-export const SelectGroup = SelectPrimitive.Group
-export const SelectValue = SelectPrimitive.Value
+// // 선택 컴포넌트
+// export const Select = SelectPrimitive.Root
+// export const SelectGroup = SelectPrimitive.Group
+// export const SelectValue = SelectPrimitive.Value
 
-export const SelectTrigger = forwardRef(({ className, children, ...props }, ref) => (
-  <SelectPrimitive.Trigger
-    ref={ref}
-    className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-    {...props}
-  >
-    {children}
-    <ChevronDown className="h-4 w-4 opacity-50" />
-  </SelectPrimitive.Trigger>
-))
-SelectTrigger.displayName = SelectPrimitive.Trigger.displayName
+// export const SelectTrigger = forwardRef(({ className, children, ...props }, ref) => (
+//   <SelectPrimitive.Trigger
+//     ref={ref}
+//     className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+//     {...props}
+//   >
+//     {children}
+//     <ChevronDown className="h-4 w-4 opacity-50" />
+//   </SelectPrimitive.Trigger>
+// ))
+// SelectTrigger.displayName = SelectPrimitive.Trigger.displayName
 
-export const SelectContent = forwardRef(({ className, children, position = "popper", ...props }, ref) => (
-  <SelectPrimitive.Portal>
-    <SelectPrimitive.Content
-      ref={ref}
-      className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
-      position={position}
-      {...props}
-    >
-      <SelectPrimitive.Viewport className="p-1">{children}</SelectPrimitive.Viewport>
-    </SelectPrimitive.Content>
-  </SelectPrimitive.Portal>
-))
-SelectContent.displayName = SelectPrimitive.Content.displayName
+// export const SelectContent = forwardRef(({ className, children, position = "popper", ...props }, ref) => (
+//   <SelectPrimitive.Portal>
+//     <SelectPrimitive.Content
+//       ref={ref}
+//       className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
+//       position={position}
+//       {...props}
+//     >
+//       <SelectPrimitive.Viewport className="p-1">{children}</SelectPrimitive.Viewport>
+//     </SelectPrimitive.Content>
+//   </SelectPrimitive.Portal>
+// ))
+// SelectContent.displayName = SelectPrimitive.Content.displayName
 
-export const SelectItem = forwardRef(({ className, children, ...props }, ref) => (
-  <SelectPrimitive.Item
-    ref={ref}
-    className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
-    {...props}
-  >
-    <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-      <SelectPrimitive.ItemIndicator>
-        <Check className="h-4 w-4" />
-      </SelectPrimitive.ItemIndicator>
-    </span>
-    <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
-  </SelectPrimitive.Item>
-))
-SelectItem.displayName = SelectPrimitive.Item.displayName
+// export const SelectItem = forwardRef(({ className, children, ...props }, ref) => (
+//   <SelectPrimitive.Item
+//     ref={ref}
+//     className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
+//     {...props}
+//   >
+//     <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
+//       <SelectPrimitive.ItemIndicator>
+//         <Check className="h-4 w-4" />
+//       </SelectPrimitive.ItemIndicator>
+//     </span>
+//     <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
+//   </SelectPrimitive.Item>
+// ))
+// SelectItem.displayName = SelectPrimitive.Item.displayName
 
 // // 대화상자 컴포넌트
 // export const Dialog = DialogPrimitive.Root
